+++ CHANGELOG.md
@@ -68,6 +68,7 @@ You can find its changes [documented below](#070---2021-01-01).
 - Fixed layout of scrollbar with very small viewports ([#1715] by [@andrewhickman])
 - Fixed `WindowLevel::Tooltip` on Windows platform ([#1737] by [@djeedai])
 - X11 backend now supports scaling([#1751] by [@Maan2003])
+- X11 backend now supports changing cursors ([#1755] by [@Maan2003])
 
 ### Visual
 
@@ -701,6 +702,7 @@ Last release without a changelog :(
 [#1737]: https://github.com/linebender/druid/pull/1737
 [#1743]: https://github.com/linebender/druid/pull/1743
 [#1751]: https://github.com/linebender/druid/pull/1751
+[#1755]: https://github.com/linebender/druid/pull/1755
 
 [Unreleased]: https://github.com/linebender/druid/compare/v0.7.0...master
 [0.7.0]: https://github.com/linebender/druid/compare/v0.6.0...v0.7.0